refactor: move update_agent_action_cost function to agent module
